requirements.txt
duckdb==1.1.3
pandas==2.2.3
python-dotenv==1.0.1
tqdm==4.67.1
pyarrow==18.1.0
